chore: move all hooks to a hooks folder
fix: lint

chore: add deprecation notices
