+++ packages/strapi-design-system/src/Field/Field.js
@@ -3,7 +3,7 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 
 import { FieldContext } from './FieldContext';
-import { useId } from '../helpers/useId';
+import { useId } from '../hooks/useId';
 
 export const Field = ({ children, name, error, hint, id, required, ...props }) => {
   const generatedId = useId(id);
