+++ packages/strapi-design-system/src/NumberInput/NumberInput.js
@@ -9,8 +9,8 @@ import styled from 'styled-components';
 import { Field, FieldLabel, FieldHint, FieldError, FieldInput } from '../Field';
 import { getDefaultLocale } from '../helpers/getDefaultLocale';
 import { KeyboardKeys } from '../helpers/keyboardKeys';
-import { useId } from '../helpers/useId';
 import { useControllableState } from '../hooks/useControllableState';
+import { useId } from '../hooks/useId';
 import { Icon } from '../Icon';
 import { Stack } from '../Stack';
 import { sizes } from '../themes/sizes';
