+++ packages/strapi-design-system/src/helpers/useId.ts
@@ -1,17 +1,13 @@
-import React, { useLayoutEffect, useState } from 'react';
+import { once, prefix } from './deprecations';
+import { useId as actualUseId } from '../hooks/useId';
 
-// Inspired by radix-ui useId hook https://github.com/radix-ui/primitives/blob/main/packages/react/id/src/id.tsx
-// We `toString()` to prevent bundlers from trying to `import { useId } from 'react';`
-const useReactId = (React as any)['useId'.toString()] || (() => undefined);
-let count = 0;
+const warnDeprecated = once(console.warn);
 
-export const useId = (initialId?: string | number | undefined): string => {
-  const [id, setId] = useState(useReactId());
+/**
+ * @deprecated useId has moved. Please import it from "@strapi/design-system/hooks/useId"
+ */
+export const useId: typeof actualUseId = (...args) => {
+  warnDeprecated(`${prefix} useId has moved. Please import it from "@strapi/design-system/hooks/useId"`);
 
-  // React versions older than 18 will have client-side ids only.
-  useLayoutEffect(() => {
-    if (!initialId) setId((reactId) => reactId ?? String(count++));
-  }, [initialId]);
-
-  return initialId?.toString() ?? (id || '');
+  return actualUseId(...args);
 };
