+++ packages/strapi-design-system/src/hooks/useLockScroll.ts
@@ -1,0 +1,15 @@
+import { useEffect } from 'react';
+
+const useLockScroll = (lockScroll: boolean) => {
+  useEffect(() => {
+    if (lockScroll) {
+      document.body.classList.add('lock-body-scroll');
+    }
+
+    return () => {
+      document.body.classList.remove('lock-body-scroll');
+    };
+  }, [lockScroll]);
+};
+
+export default useLockScroll;
