+++ packages/strapi-design-system/src/hooks/usePrevious.ts
@@ -1,0 +1,10 @@
+import { useRef, useEffect } from 'react';
+
+export const usePrevious = <TValue>(value: TValue): TValue | undefined => {
+  const ref = useRef<TValue>();
+  useEffect(() => {
+    ref.current = value;
+  });
+
+  return ref.current;
+};
