+++ packages/strapi-design-system/src/index.ts
@@ -1,3 +1,6 @@
+/* -------------------------------------------------------------------------------------------------
+ * Components
+ * -----------------------------------------------------------------------------------------------*/
 export * from './Accordion';
 export * from './Alert';
 export * from './Avatar';
