+++ packages/strapi-design-system/src/v2/SubNav/SubNavHeader.js
@@ -8,8 +8,8 @@ import { Box } from '../../Box';
 import { Divider } from '../../Divider';
 import { Flex } from '../../Flex';
 import { KeyboardKeys } from '../../helpers/keyboardKeys';
-import { useId } from '../../helpers/useId';
-import { usePrevious } from '../../helpers/usePrevious';
+import { useId } from '../../hooks/useId';
+import { usePrevious } from '../../hooks/usePrevious';
 import { IconButton } from '../../IconButton';
 import { Searchbar, SearchForm } from '../../Searchbar';
 import { Typography } from '../../Typography';
